Move Hash.cpp/hpp from core to shared_core
